+++ packages/jira-adapter/src/filters/workflow/transition_structure.ts
@@ -20,18 +20,18 @@ import _ from 'lodash'
 import { Status, Transition, WorkflowInstance } from './types'
 import { SCRIPT_RUNNER_POST_FUNCTION_TYPE } from '../script_runner/workflow/workflow_cloud'
 
-const PARTS_SEPARATOR = '::'
-
-const TYPE_TO_FROM_MAP = new Map([
-  ['Initial', 'none'],
-  ['Global', 'any status'],
-  ['Circular', 'any status'],
-])
+export const TRANSITION_PARTS_SEPARATOR = '::'
 
 type TransitionType = 'Directed' | 'Initial' | 'Global' | 'Circular'
 
+const TYPE_TO_FROM_MAP: Record<Exclude<TransitionType, 'Directed'>, string> = {
+  Initial: 'none',
+  Global: 'any status',
+  Circular: 'any status',
+}
+
 const getTransitionTypeFromKey = (key: string): string => {
-  const type = invertNaclCase(key).split(PARTS_SEPARATOR).pop() as string
+  const type = invertNaclCase(key).split(TRANSITION_PARTS_SEPARATOR).pop() as string
   return type === 'Circular' ? 'Global' : type
 }
 
@@ -83,18 +83,14 @@ export const getTransitionKey = (transition: Transition, statusesMap: Map<string
       .map(from => statusesMap.get(from) ?? from)
       .sort()
       .join(',')
-    : TYPE_TO_FROM_MAP.get(type) ?? ''
+    : TYPE_TO_FROM_MAP[type]
 
-  return naclCase([transition.name ?? '', `From: ${fromSorted}`, type].join(PARTS_SEPARATOR))
+  return naclCase([transition.name ?? '', `From: ${fromSorted}`, type].join(TRANSITION_PARTS_SEPARATOR))
 }
 
 export const transformTransitions = (value: Value): SaltoError[] => {
   const statusesMap = createStatusMap(value.statuses ?? [])
-  const maxCounts: Record<string, number> = {}
-  value.transitions.forEach((transition: Transition) => {
-    const key = getTransitionKey(transition, statusesMap)
-    maxCounts[key] = (maxCounts[key] ?? 0) + 1
-  })
+  const maxCounts = _(value.transitions).map(transition => getTransitionKey(transition, statusesMap)).countBy().value()
 
   const counts: Record<string, number> = {}
 
@@ -104,18 +100,18 @@ export const transformTransitions = (value: Value): SaltoError[] => {
       const key = getTransitionKey(transition, statusesMap)
       counts[key] = (counts[key] ?? 0) + 1
       if (maxCounts[key] > 1) {
-        return [naclCase(`${invertNaclCase(key)}${PARTS_SEPARATOR}${counts[key]}`), transition]
+        return [naclCase(`${invertNaclCase(key)}${TRANSITION_PARTS_SEPARATOR}${counts[key]}`), transition]
       }
       return [key, transition]
     }))
-  const errorKeys = Object.entries(counts)
+  const errorKeyNames = Object.entries(counts)
     .filter(([, count]) => count > 1)
-    .map(([key]) => key)
+    .map(([key]) => invertNaclCase(key).split(TRANSITION_PARTS_SEPARATOR)[0])
 
-  return errorKeys.length === 0
+  return errorKeyNames.length === 0
     ? []
     : [{
-      message: `The following transitions of workflow ${value.name} have the same key: ${errorKeys.join(', ')}.
+      message: `The following transitions of workflow ${value.name} are not unique: ${errorKeyNames.join(', ')}.
 It is strongly recommended to rename these transitions so they are unique in Jira, then re-fetch`,
       severity: 'Warning',
     }]
